A single cron running all sscripts sequentially Fixes #10
When 4 scripts are set to run in a cron, they are run in parallel. That very
factor gave rise to so many troubles. Instead of doing that, why not run those
scripts in sequence. This might make a script that runs for long but it would
relieve us from the critical section problem once and for all.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>