Logging integration in view #7
Only small debug messages containing function names is added. As this file is
mainly for printing the messages to the screen, I felt logging is not much
needed.

Also, not to distort the script output, the starting and finished logging
messages has been changed from level INFO to level DEBUG.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>