Log view improvements
Using ccze utility, the logs are exported to html format which is viewable on
the web. README updated to include this dependency.

The bash script is modified to include shopt option in order to enable
bash_aliases expansion which happens only on interactive mode.

Time zone is converted to Asia/Kolkata in order to make referencing of logs
easier.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>